refactor: split out functions to seperate files
- Add Makefile
- Add header files
- Move files to sub dirs
