+++ Makefile
@@ -1,0 +1,22 @@
+CXX       := g++
+CXX_FLAGS := -std=c++17 -ggdb
+
+BIN     := bin
+SRC     := src
+INCLUDE := include
+
+LIBRARIES   :=
+EXECUTABLE  := periodic
+
+
+all: $(BIN)/$(EXECUTABLE)
+
+run: clean all
+	clear
+	./$(BIN)/$(EXECUTABLE)
+
+$(BIN)/$(EXECUTABLE): $(SRC)/*.c
+	$(CXX) $(CXX_FLAGS) -I$(INCLUDE) $^ -o $@ $(LIBRARIES)
+
+clean:
+	-rm $(BIN)/*
